harness: absorb semantic gate into providers
